+++ vscode-extensions/vscode-concourse/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2022-04-27 (4.14.1 RELEASE, incl. language servers version 1.33.0)
+
+* no major changes
+
 ## 2022-03-16 (4.14.0 RELEASE, incl. language servers version 1.32.0)
 
 * no major changes